06_AI_chain/pyproject.toml
[project]
name = "06-ai-chain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.33.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-huggingface>=0.3.0",
    "langchain-ollama>=0.3.3",
    "python-dotenv>=1.1.0",
    "transformers>=4.52.4",
]
